requirements.txt
Flask==2.1.2
numpy==1.24.1
tensorflow==2.13.0
Pillow==10.1.0
matplotlib==3.7.1
gunicorn==20.1.0
